Less intrusive changes
