+++ gpu_environment.py
@@ -1,6 +1,5 @@
 # coding=utf-8
-# Copyright 2018 The Google AI Language Team Authors.
-# Copyright (c) 2018, NVIDIA CORPORATION.  All rights reserved.
+# Copyright 2018 The TensorFlow Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,10 +13,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import fp16_utils
 import tensorflow as tf
+import numpy as np
 
 FLAGS = tf.flags.FLAGS
 
-custom_getter = fp16_utils.float32_variable_storage_getter if FLAGS.use_fp16 else None
+def float32_variable_storage_getter(getter, name, shape=None, dtype=None,
+                                    initializer=None, regularizer=None,
+                                    trainable=True,
+                                    *args, **kwargs):
+    """Custom variable getter that forces trainable variables to be stored in
+       float32 precision and then casts them to the training precision.
+    """
+    storage_dtype = tf.float32 if trainable else dtype
+    variable = getter(name, shape, dtype=storage_dtype,
+                      initializer=initializer, regularizer=regularizer,
+                      trainable=trainable,
+                      *args, **kwargs)
+    if trainable and dtype != tf.float32:
+        variable = tf.cast(variable, dtype)
+    return variable
+
+custom_getter = float32_variable_storage_getter if FLAGS.use_fp16 else None
 compute_type = tf.float16 if FLAGS.use_fp16 else tf.float32
